// src/module/audit/enums/audit-action.enum.ts
export enum AuditAction {
  USER_REGISTERED = 'USER_REGISTERED',
  EMAIL_VERIFICATION_SENT = 'EMAIL_VERIFICATION_SENT',
  EMAIL_VERIFIED = 'EMAIL_VERIFIED',
  VERIFICATION_EMAIL_RESENT = 'VERIFICATION_EMAIL_RESENT',
  USER_LOGIN = 'USER_LOGIN',
  USER_LOGOUT = 'USER_LOGOUT',
  FAILED_LOGIN = 'FAILED_LOGIN',
  PASSWORD_RESET_REQUESTED = 'PASSWORD_RESET_REQUESTED',
  PASSWORD_RESET_COMPLETED = 'PASSWORD_RESET_COMPLETED',
  PASSWORD_RESET_TOKEN_INVALID = 'PASSWORD_RESET_TOKEN_INVALID',
  PASSWORD_RESET_TOKEN_EXPIRED = 'PASSWORD_RESET_TOKEN_EXPIRED',
  WORKSPACE_INVITATION_CREATED = 'WORKSPACE_INVITATION_CREATED',
  WORKSPACE_INVITATION_ACCEPTED = 'WORKSPACE_INVITATION_ACCEPTED',
  WORKSPACE_INVITATION_DECLINED = 'WORKSPACE_INVITATION_DECLINED',
  WORKSPACE_INVITATION_CANCELLED = 'WORKSPACE_INVITATION_CANCELLED',
  GOOGLE_LOGIN = 'GOOGLE_LOGIN',
  GOOGLE_ACCOUNT_CREATED = 'GOOGLE_ACCOUNT_CREATED',
  GOOGLE_ACCOUNT_LINKED = 'GOOGLE_ACCOUNT_LINKED',
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  TRASH_ITEM_RESTORED = 'TRASH_ITEM_RESTORED',
  TRASH_EMPTIED = 'TRASH_EMPTIED',
}